[781] Move to event mapping engine
