+++ app/services/event_mapping_engine/map/historical.rb
@@ -1,0 +1,12 @@
+module EventMappingEngine
+  module Map
+    class Historical
+      def initialize
+      end
+
+      def run
+        # IMPLEMENT
+      end
+    end
+  end
+end
